Re-order points of the compass
I found it hard thinking about "left" meaning a change in the index into
the list of compass points of +1 and "right" meaning a change of -1.

Now that the code is complete with a full suite of tests, it feels safe
to fix that problem by re-ordering the list so that left and right make
sense, intuitively.