Remove conditional that handles 'F' command
Are you wondering what the hell's going on with this commit?

The motivation for this change was to remove an `if` statement. As a
rule of thumb, code without conditional statements is cheaper to
maintain.

The `lambda` keyword defines anonymous functions, so what we're doing
here is:

- defining the default behaviour for instructions passed to `.move()` as
  `self._travel(command, argument)`

- defining custom behaviour for the L, R and F commands.

As a result, the `_travel()` method no longer needs to know that "F" is
a special case, and the fact that "F" maps to the current value of
`self.facing` has become data. It's still hard-coded, but that code has
slightly easier to extend with new behaviour.

That's the point of Bertrand Meyer's "open-closed principle" [1], one of
the five principles that make up SOLID [2].

[1] https://en.wikipedia.org/wiki/Open%E2%80%93closed_principle
[2] https://en.wikipedia.org/wiki/SOLID